Add the initial package support for the Review Board installer.
The Review Board installer is the soon-to-be-recommended way of
installing Review Board, supporting a wide range of Linux distributions
and versions.

This initial change sets up the repository and the Python package,
preparing for the beginnings of the installer code.

Testing Done:
Built the (mostly empty) package successfully.

Tested this with the upcoming changes to introduce a full installer.

Reviewed at https://reviews.reviewboard.org/r/13460/