Add support for installing packages using a variety of methods.
This introduces an abstraction layer for installing packages using a
variety of methods. This includes:

* `apt-get build-dep`
* `apt-get install`
* `brew install`
* `curl <url> | pythonX.Y`
* `pacman`
* `pip install ReviewBoard[...]`
* `pip install`
* `yum install`
* `zypper install`
* Arbitrary shell commands

This is all done by running `run_install_method()` and passing in the
appropriate arguments. It will be used for package installation in an
upcoming change.

Testing Done:
Unit tests pass.

Tested each of these with the rest of the upcoming installer changes.

Reviewed at https://reviews.reviewboard.org/r/13464/