Add functionality for querying PyPI for packages.
This introduces `rbinstall.pypi`, which uses the PyPI JSON API to look
up information on packages and determine which are available for install
on the target version of Python, based on any requested package
versions.

This will be used to notify users in the installer as to which version
of Python their system is compatible with, and whether that's the latest
version released.

Testing Done:
Unit tests pass.

Tested this with the rest of the installer code.

Reviewed at https://reviews.reviewboard.org/r/13466/