Add the main entry point for the installer.
This adds a simple `main()` for performing the installation. It
scans and hard-codes some installer settings, determines the latest
version of Review Board, and then begins the installation.

This is kept pretty basic for now. Future changes will build on this to
guide the user and to allow for some customization of the install
process.

Testing Done:
Tested the installer in all supported environments.

Reviewed at https://reviews.reviewboard.org/r/13467/