Add scripts for building a bootstrap installer.
In order to simplify installation, we'll be distributing the Review
Board installer as a single file, able to be run using `curl | python`.

To do this, we'll be distributing the primary installer as a
booststrapping Python script that contains an embedded copy of the
installer's wheel. The bootstrapping installer sets up a temporary
virtualenv (using `python -m venv`), installs the package within that,
and then runs it. This ensures we can run in a stable environment.

`scripts/pack.sh` is responsible for building the bootstrap installer
and the wheel, and then embedding the wheel within the bootstrap
installer.

Testing Done:
Successfully built the installer and ran it in all supported Linux
distributions.

Reviewed at https://reviews.reviewboard.org/r/13468/