Fix issues bootstrapping the installer.
When bootstrapping the installer, the bootstrapper failed to pass on
stdin to the installer program. This had worked previously, but was
broken when fixing another issue. We now pass in the stdin file handle,
allowing the installer to open stdin for input.

There's also a fix to remove the pip and Python version checks when
installing the bootstrapper package, avoiding a warning that shouldn't
be present during the bootstrapping process.

Testing Done:
Successfully bootstrapped on macOS and Linux, without hitting issues
with stdin.

No longer saw the Python/pip version warnings on an older distro.

Reviewed at https://reviews.reviewboard.org/r/13504/