Don't attempt an implicit unattended install when outputting to a file.
When running the installer with output piped to a file (or otherwise not
writing to a terminal), the installer would accept all defaults for
prompts and then initiate the install. This may not be what people want.

We now require `--noinput` in order to do an unattended install. When
outputting to a non-terminal without passing `--noinput`, the installer
will exit with an error containing instructions.

Testing Done:
Tested the results when piping to a file.

Tested default attended installs and unattended with `--noinput`.

Reviewed at https://reviews.reviewboard.org/r/13506/