Fix issues working with stdin when piping to Python.
We hit an issue before where piping the installer to Python would give
us a useless stdin. I had thought this was fixed, but I've still hit
errors with the new code here.

First off, in the cases I've tested locally, stdin isn't coming up as a
TTY, which just means none of this worked. In my past tests, it did.
This might be dependent on Python in some way. I'm not sure.

The solution here is to re-open stdin as a TTY. This gives us back full
input from the terminal, and seems to be the standard practice (though
time will tell if there are gotchas).

Second, querying for terminal info could fail. We now check for
exceptions here and just fall back to an empty result.

Testing Done:
Tested running the installer manually, via `curl | python`, and via
`cat | python` on both macOS and Linux (over SSH). Verified I was able
to run the installer, provide input that the installer could work with,
and that the terminal queries continued to work (including detecting light
backgrounds).

Reviewed at https://reviews.reviewboard.org/r/13555/