Add instructions and options for running the installer in a safe mode.
The installer now respects `RBINSTALL_COLOR` and `RBINSTALL_INTERACTIVE`
environment variables. The former allows for the equivalent of
`--no-color` (when set to `0`), and the latter ``--noinput` (when set
to `0`).

The bootstrapper now provides some safety instructions immediately
before handling control off to the main installer. These tell the user
that if they can't see or run the installer, set `RBINSTALL_COLOR=0` and
try again, or contact support.

This is intended as an aid for anyone who may have trouble seeing the
installer. One user encountered this in a tmux environment running on
Ubuntu Linux in, presumably, GNOME Terminal. While we haven't been able
to reproduce the cause, the hope is that this will prevent people from
getting completely stuck.

Testing Done:
Ran the installer and saw the new bootstrap text changes.

Verified that the new options controlled the behavior of the installer.

Reviewed at https://reviews.reviewboard.org/r/13685/