Don't query terminal colors if we're disabling color support.
When starting up, we check the terminal background for light or dark
colors. This happens even if we've set no-color mode. That goes against
the intent of the workaround for terminal issues.

We no longer perform this check if color isn't allowed.

Testing Done:
Ran in no-color mode and verified that a terminal query was never
performed.

Reviewed at https://reviews.reviewboard.org/r/14250/